# swagger/center.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "OKR 系统API",
    "description": "OKR 系统相关接口文档！",
    "contact": {
      "name": "发际线与我作对",
      "url": "https://github.com/CarefreeState/OKR-System",
      "email": "2040484356@qq.com"
    },
    "version": "3.0"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:1701",
      "description": "Inferred Url"
    }
  ],
...
  "paths": {
...
    "/keyresult/add": {
      "post": {
        "tags": [
          "关键结果"
        ],
        "summary": "添加关键结果",
        "operationId": "addKeyResultUsingPOST",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/添加关键结果所需数据"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/统一响应"
                }
              }
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/keyresult/update": {
      "post": {
        "tags": [
          "关键结果"
        ],
        "summary": "更新完成概率",
        "operationId": "updateKeyResultUsingPOST",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/更新关键结果所需数据"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/统一响应"
                }
              }
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
    "/flag/add": {
      "post": {
        "tags": [
          "状态指标"
        ],
        "summary": "增加一条状态指标",
        "operationId": "addStatusFlagUsingPOST",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/添加状态指标所需数据"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/统一响应"
                }
              }
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/flag/remove": {
      "post": {
        "tags": [
          "状态指标"
        ],
        "summary": "删除一条指标",
        "operationId": "removeUsingPOST",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "description": "状态指标 ID",
            "required": false,
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "scene",
            "in": "query",
            "description": "场景",
            "required": false,
            "style": "form",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/统一响应"
                }
              }
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/flag/update": {
      "post": {
        "tags": [
          "状态指标"
        ],
        "summary": "更新一条指标",
        "operationId": "updateUsingPOST",
        "parameters": [
          {
            "name": "scene",
            "in": "query",
            "required": false,
            "style": "form",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "statusFlagUpdateDTO.color",
            "in": "query",
            "description": "颜色",
            "required": false,
            "style": "form",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "statusFlagUpdateDTO.id",
            "in": "query",
            "description": "指标 ID",
            "required": false,
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "statusFlagUpdateDTO.label",
            "in": "query",
            "description": "指标内容",
            "required": false,
            "style": "form",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/统一响应"
                }
              }
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "关键结果数据": {
        "title": "关键结果数据",
        "type": "object",
        "properties": {
          "content": {
            "type": "string",
            "description": "关键结果内容"
          },
          "firstQuadrantId": {
            "type": "integer",
            "description": "第一象限 ID",
            "format": "int64"
          },
          "probability": {
            "type": "integer",
            "description": "完成概率",
            "format": "int32"
          }
        }
      },
      "关键结果更新数据": {
        "title": "关键结果更新数据",
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "关键结果 ID",
            "format": "int64"
          },
          "probability": {
            "type": "integer",
            "description": "完成概率",
            "format": "int32"
          }
        }
      },
...
      "更新关键结果所需数据": {
        "title": "更新关键结果所需数据",
        "type": "object",
        "properties": {
          "keyResultUpdateDTO": {
            "description": "更新关键结果的数据",
            "$ref": "#/components/schemas/关键结果更新数据"
          },
          "scene": {
            "type": "string",
            "description": "场景"
          }
        }
      },
      "添加关键结果所需数据": {
        "title": "添加关键结果所需数据",
        "type": "object",
        "properties": {
          "keyResultDTO": {
            "description": "关键结果数据",
            "$ref": "#/components/schemas/关键结果数据"
          },
          "scene": {
            "type": "string",
            "description": "场景"
          }
        }
      },
      "添加状态指标所需数据": {
        "title": "添加状态指标所需数据",
        "type": "object",
        "properties": {
          "scene": {
            "type": "string",
            "description": "场景"
          },
          "statusFlagDTO": {
            "description": "状态指标数据",
            "$ref": "#/components/schemas/状态指标数据"
          }
        }
      },
...
      "状态指标数据": {
        "title": "状态指标数据",
        "type": "object",
        "properties": {
          "color": {
            "type": "string",
            "description": "颜色"
          },
          "fourthQuadrantId": {
            "type": "integer",
            "description": "第四象限 ID",
            "format": "int64"
          },
          "label": {
            "type": "string",
            "description": "指标内容"
          }
        }
      },
...
      "统一响应": {
        "title": "统一响应",
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "状态码",
            "format": "int32"
          },
          "data": {
            "type": "object",
            "description": "具体数据"
          },
          "message": {
            "type": "string",
            "description": "描述语"
          }
        }
      },
...
    }
  }
}